+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #
 - Add `ContainerCreateOptsBuilder::network_config`
+- `Docker` initializers like `new`, `tcp`, `unix`, `tls` now create an unversioned connector that will use the server's latest version instead of setting it to `LATEST_API_VERSION`.
+  This means that by default this crate will be easier to use with older versions of Docker.
 
 # 0.13.0
 - Fix Container::attach output when TTY is enabled on container